Move DBErrorDetailsProvider to io.cdap.plugin.common.db
